# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shops-owners": {
      "get": {
        "description": "Возвращает список всех владельцев магазинов с информацией об их магазинах. Требует роль ROOT или MANAGER.",
        "operationId": "ShopOwnersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список владельцев успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                      "name": "Иванов Иван Иванович",
                      "phone": "+79991234567",
                      "email": "ivanov@example.com",
                      "address": "г. Москва, ул. Пушкина, д. 5, кв. 10",
                      "inn": "771234567890",
                      "createdAt": "2025-10-19T10:00:00.000Z",
                      "updatedAt": "2025-10-19T10:00:00.000Z",
                      "shops": [
                        {
                          "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                          "name": "Магазин \"Продукты 24\"",
                          "address": "г. Москва, ул. Ленина, д. 10",
                          "login": "shop_products24"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить список всех владельцев магазинов",
        "tags": [
          "Владельцы магазинов"
        ]
      },
      "post": {
        "description": "Создание нового владельца магазинов в системе. Обязательные поля: имя и телефон. Email, адрес и ИНН необязательны.",
        "operationId": "ShopOwnersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopOwnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Владелец успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                    "name": "Иванов Иван Иванович",
                    "phone": "+79991234567",
                    "email": "ivanov@example.com",
                    "address": "г. Москва, ул. Пушкина, д. 5, кв. 10",
                    "inn": "771234567890",
                    "createdAt": "2025-10-19T10:00:00.000Z",
                    "updatedAt": "2025-10-19T10:00:00.000Z",
                    "shops": []
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные (например, неверный формат телефона или email)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "phone must be a valid phone number",
                      "email must be an email"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Создать нового владельца магазинов",
        "tags": [
          "Владельцы магазинов"
        ]
      }
    },
    "/shops-owners/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о конкретном владельце магазина, включая список всех его магазинов.",
        "operationId": "ShopOwnersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о владельце успешно получена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                    "name": "Иванов Иван Иванович",
                    "phone": "+79991234567",
                    "email": "ivanov@example.com",
                    "address": "г. Москва, ул. Пушкина, д. 5, кв. 10",
                    "inn": "771234567890",
                    "createdAt": "2025-10-19T10:00:00.000Z",
                    "updatedAt": "2025-10-19T10:00:00.000Z",
                    "shops": [
                      {
                        "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                        "name": "Магазин \"Продукты 24\"",
                        "address": "г. Москва, ул. Ленина, д. 10",
                        "inn": "7707083893",
                        "login": "shop_products24"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Владелец не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop owner not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить информацию о владельце по ID",
        "tags": [
          "Владельцы магазинов"
        ]
      },
      "patch": {
        "description": "Обновление данных владельца магазинов (имя, телефон, email, адрес, ИНН). Все поля опциональны при обновлении.",
        "operationId": "ShopOwnersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShopOwnerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Информация о владельце успешно обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                    "name": "Иванов Иван Петрович",
                    "phone": "+79991234567",
                    "email": "ivanov.new@example.com",
                    "address": "г. Москва, ул. Пушкина, д. 5, кв. 10",
                    "inn": "771234567890",
                    "createdAt": "2025-10-19T10:00:00.000Z",
                    "updatedAt": "2025-10-21T15:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "email must be an email"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Владелец не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop owner not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Обновить информацию о владельце",
        "tags": [
          "Владельцы магазинов"
        ]
      },
      "delete": {
        "description": "Удаление владельца из системы. ВНИМАНИЕ: При удалении владельца все связанные с ним магазины также будут удалены (каскадное удаление). Терминалы магазинов не удаляются, но теряют связь с магазином.",
        "operationId": "ShopOwnersController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Владелец успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop owner deleted successfully",
                    "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Владелец не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop owner not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Удалить владельца магазинов",
        "tags": [
          "Владельцы магазинов"
        ]
      }
    },
...
    "/shops/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о конкретном магазине, включая данные владельца и список терминалов.",
        "operationId": "ShopsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о магазине успешно получена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3",
                    "name": "Магазин \"Продукты 24\"",
                    "address": "г. Москва, ул. Ленина, д. 10",
                    "inn": "7707083893",
                    "kpp": "770701001",
                    "ogrn": "1027700132195",
                    "login": "shop_products24",
                    "ownerId": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                    "createdAt": "2025-10-20T10:00:00.000Z",
                    "updatedAt": "2025-10-20T10:00:00.000Z",
                    "owner": {
                      "id": "f5b9e8a1-3c4d-4e5f-9a8b-7c6d5e4f3a2b",
                      "name": "Иванов Иван Иванович",
                      "phone": "+79991234567",
                      "email": "ivanov@example.com"
                    },
                    "terminals": []
                  }
                }
              }
            }
          },
          "404": {
            "description": "Магазин не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить информацию о магазине по ID",
        "tags": [
          "Магазины"
        ]
      },
      "delete": {
        "description": "Удаление магазина из системы. При удалении магазина все привязанные к нему терминалы теряют связь с магазином (shopId устанавливается в null), но сами терминалы не удаляются.",
        "operationId": "ShopsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Магазин успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop deleted successfully",
                    "id": "a96e7bad-d4d8-4559-b9b7-0e180fedd3e3"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Магазин не найден",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Shop not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Удалить магазин",
        "tags": [
          "Магазины"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shop Terminals Management API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен полученный при авторизации",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateShopOwnerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ИП Иванов Иван Иванович",
            "description": "Полное наименование владельца магазина (ФИО или название организации)"
          },
          "phone": {
            "type": "string",
            "example": "+79001234567",
            "description": "Контактный телефон"
          },
          "email": {
            "type": "string",
            "example": "ivanov@example.com",
            "description": "Email для связи"
          },
          "address": {
            "type": "string",
            "example": "г. Ростов-на-Донц, ул. Соколова 76",
            "description": "Адрес регистрации владельца"
          },
          "inn": {
            "type": "string",
            "example": "123456789012",
            "description": "ИНН (для юридических лиц и ИП)"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "UpdateShopOwnerDto": {
        "type": "object",
        "properties": {

        }
      },
...
    }
  }
}